wip
All compiler functions to CompilerContext, and fixes. Support Windows/Mac/Linux in source roots by normalising path sep. to Mac/Linux
